pyproject.toml
[tool.poetry]
name = "Sudoku API"
version = "0.1.0"
description = "Sudoku API with FastAPI and PySudoku"
authors = ["hasansezertasan <hasansezertasan@gmail.com>"]
readme = "README.md"
maintainers = ["hasansezertasan <hasansezertasan@gmail.com>"]
packages = [{ include = "toolbox.py" }]
license = "MIT"


[tool.ruff]
# Run `ruff linter` command to see all available linters and their descriptions
target-version = "py311"
line-length = 120
select = [
  # "B",
  # "C",
  "E",
  "F",
  "W",
  "I",
  "B9",
]
ignore = [
  "B008", # do not perform function calls in argument defaults
  "C901", # too complex
  "E501", # line too long, handled by black
  "F401", # imported but unused
  "E402", # module level import not at top of file
  "E712", # comparison to True should be 'if cond is True:' or 'if cond:'
]
exclude = [
  "alembic",
  "migrations",
]
show-fixes = true

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
split_on_trailing_comma = true
skip_glob = ["**/alembic/**", "**/migrations/**"]


[tool.black]
line-length = 120
target-version = ['py311']
exclude = '''
(
  /(
      \.eggs
    | \.git
    | \.hg
    | \.mypy_cache
    | \.nox
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
    | alembic
  )/
)
'''


[tool.poetry.scripts]
toolbox = "toolbox:app"

[tool.poe.tasks]
line-count = "pygount --format=summary"
test = "pytest -v -s"
black = "black ."
isort = "isort ."
ruff = "ruff ."
bir = ["black", "isort", "ruff"]
run = "uvicorn --port 5000 --reload main:app"
toolbox = { script = "toolbox.__main__:app" }
clean = "frenchmaid clean"
requirements = "poetry export -f requirements.txt --output pyproject.txt --without-hashes"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
python-dotenv = "^1.0.0"
py-sudoku = "^1.0.3"
typer = { extras = ["all"], version = "^0.9.0" }
uvicorn = "0.23.2"
fastapi = "^0.103.1"
pydantic = "2.3.0"
python-multipart = "~0.0.6"
email-validator = "^2.0.0.post2"
httpx = "^0.24.1"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.25.1"
isort = "^5.12.0"
black = { extras = ["jupyter"], version = "^23.7.0" }
ruff = "^0.0.287"
pre-commit = "^3.4.0"
poethepoet = "^0.22.0"
frenchmaid = "^0.2.3"

[tool.poetry.group.test.dependencies]
pytest = "^7.4.0"

# poetry install --with docs
[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
pdoc3 = "^0.10.0"

[tool.poetry.group.analysis]
optional = true

[tool.poetry.group.analysis.dependencies]
pygount = "1.6.1"
